perf(008): stream every HTTP blob-read path (no full-buffer), share one streaming helper
The by-id read paths (ServeDocument public serve, GetContent internal read) buffered the WHOLE blob
into a []byte via Storage.Read before writing it — so a burst of concurrent large-file reads could
drive RSS to N×blobsize and OOM the pod. The by-hash GetBlobContent already streamed; now all three
share one helper.

- New streamBlob (response.go): copies rc→w in constant memory after the caller writes status +
  Content-Length, and classifies a mid-stream failure once (backend read fault / short-read
  truncation → WARN + panic(http.ErrAbortHandler) so a truncated body can't pass for a clean EOF
  against the declared Content-Length; client write failure → silent). GetBlobContent's inline
  copy+switch is refactored onto it (single definition, no drift).
- ServeDocument + GetContent now ReadStream + streamBlob, adding an explicit Content-Length; a
  stored key stays a 500 on a malformed value (via writeStorageReadError), never a client 400.
- mime-repair (boot sweep) read the whole blob only to check non-empty + the 2-byte zip magic; it
  now reads a bounded 512-byte prefix over a stream (readBlobPrefix), so a large office doc can't
  spike RSS during the repair.

openapi unchanged; full suite + lint green. Remaining full-blob reads: ServeFile (dead — no caller)
and MeasureDims (image-dims metadata) — addressed separately.
